requirements.txt
pytz==2022.1
Telethon==1.28.5
requests==2.31.0
aiohttp==3.8.3
pyOpenSSL==23.1.1
